+++ src/components/ui-custom/index.stories.js
@@ -1,0 +1,10 @@
+import { define } from 'hybrids';
+import Custom from './';
+
+define('ui-custom', Custom);
+
+export default {
+  title: 'UI|<ui-custom>'
+};
+
+export const demo = () => document.createElement('ui-custom');