+++ src/components/ui-mix/index.js
@@ -3,6 +3,9 @@ import { color, typography } from '@quarksuite/core';
 import colorKnob from '../knobs/color';
 import rangeKnob from '../knobs/range';
 
+import baseStyles from '../shared/base.pcss';
+import formStyles from '../shared/form.pcss';
+
 export default {
   base: '#ff0000',
   blend: '#0000ff',
@@ -11,11 +14,15 @@ export default {
   output: ({ base, blend, amount }) => color.mix(base, blend, amount),
   render: ({ base, blend, amount, format, output }) =>
     html`
+      <style>
+        .color {
+          flex: 1;
+        }
+      </style>
       <form action="">
         ${colorKnob('Base Color', base, html.set('base'))}
         ${colorKnob('Blend Target', blend, html.set('blend'))}
         ${rangeKnob('Blend Amount', amount, html.set('amount'))}
       </form>
-      <b-swatch value="${color.format(output, format)}"></b-swatch>
-    `
+    `.style(baseStyles, formStyles)
 };