+++ src/components/ui-scheme/index.js
@@ -6,6 +6,9 @@ import rangeKnob from '../knobs/range';
 import radioKnob from '../knobs/radio';
 import checkboxKnob from '../knobs/checkbox';
 
+import baseStyles from '../shared/base.pcss';
+import formStyles from '../shared/form.pcss';
+
 const setScheme = (host, event) => (host.type = event.target.value);
 const setFormat = (host, event) => (host.format = event.target.value);
 
@@ -85,5 +88,5 @@ export default {
             `
         )}
       </div>
-    `
+    `.style(baseStyles, formStyles)
 };